+++ resources/views/components/favicon.blade.php
@@ -1,0 +1,1 @@
+<link {{ $attributes->merge(array_filter(['rel' => 'icon', 'href' => $url, 'type' => $type])) }} />