+++ resources/views/components/logo.blade.php
@@ -1,0 +1,1 @@
+<img {{ $attributes->merge(['src' => $url, 'alt' => $alt]) }} />